Fix some bugs from refactor & add dynamic feature module to prove stubs
