+++ annotations/src/main/java/com/juullabs/exercise/annotations/AsStub.kt
@@ -2,5 +2,5 @@ package com.juullabs.exercise.annotations
 
 @MustBeDocumented
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
+@Retention(AnnotationRetention.BINARY)
 annotation class AsStub(val packageName: String, val className: String)
