+++ annotations/src/main/java/com/juullabs/exercise/annotations/Exercise.kt
@@ -2,5 +2,5 @@ package com.juullabs.exercise.annotations
 
 @MustBeDocumented
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
+@Retention(AnnotationRetention.BINARY)
 annotation class Exercise(vararg val params: ExerciseParameter)
