+++ annotations/src/main/java/com/juullabs/exercise/annotations/ResultContract.kt
@@ -2,5 +2,5 @@ package com.juullabs.exercise.annotations
 
 @MustBeDocumented
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
+@Retention(AnnotationRetention.BINARY)
 annotation class ResultContract(vararg val kinds: ResultKind)
