+++ annotations/src/main/java/com/juullabs/exercise/annotations/ResultKind.kt
@@ -2,5 +2,5 @@ package com.juullabs.exercise.annotations
 
 @MustBeDocumented
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.SOURCE)
+@Retention(AnnotationRetention.BINARY)
 annotation class ResultKind(val name: String, vararg val params: ExerciseParameter)
