+++ application/build.gradle
@@ -43,17 +43,19 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
+    dynamicFeatures = [':feature']
 }
 
 dependencies {
-    implementation project(":annotations")
+    api project(":annotations")
+    api "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    api 'androidx.core:core-ktx:1.2.0'
+    api 'androidx.appcompat:appcompat:1.1.0'
+    api 'androidx.activity:activity-ktx:1.2.0-alpha02'
+    api 'androidx.fragment:fragment-ktx:1.3.0-alpha02'
     kapt project(":compile")
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.2.0'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.activity:activity-ktx:1.2.0-alpha02'
-    implementation 'androidx.fragment:fragment-ktx:1.3.0-alpha02'
     testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.annotation:annotation:1.1.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
