+++ feature/src/main/AndroidManifest.xml
@@ -1,0 +1,20 @@
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:dist="http://schemas.android.com/apk/distribution"
+    package="com.juullabs.dynamicfeature"
+    >
+
+    <dist:module
+        dist:instant="false"
+        dist:title="@string/title_dynamic_feature"
+        >
+        <dist:delivery>
+            <dist:install-time />
+        </dist:delivery>
+        <dist:fusing dist:include="false" />
+    </dist:module>
+
+    <application>
+        <activity android:name=".CalculatorChildActivity" />
+    </application>
+</manifest>
