+++ feature/src/main/java/com/juullabs/dynamicfeature/CalculatorChildActivity.kt
@@ -1,19 +1,17 @@
-package com.juullabs.exercise.activities
+package com.juullabs.dynamicfeature
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.juullabs.exercise.R
+import com.juullabs.exercise.activities.CalculatorChildActivityStub
 import com.juullabs.exercise.annotations.Exercise
-import com.juullabs.exercise.annotations.Extra
-import com.juullabs.exercise.annotations.ResultContract
-import com.juullabs.exercise.annotations.ResultKind
+import com.juullabs.exercise.annotations.FromStub
 import kotlinx.android.synthetic.main.activity_calc_child.cancel
 import kotlinx.android.synthetic.main.activity_calc_child.inputTextView
 import kotlinx.android.synthetic.main.activity_calc_child.plus1
 import kotlinx.android.synthetic.main.activity_calc_child.times3
 
-@Exercise(Extra("input", Int::class))
-@ResultContract(ResultKind("Ok", Extra("output", Int::class)))
+@Exercise
+@FromStub(CalculatorChildActivityStub::class)
 class CalculatorChildActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
