+++ feature/src/main/res/layout/activity_calc_child.xml
