+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':annotations'
 include ':application'
 include ':compile'
+include ':feature'
